[HELPS-1008] Test refactoring
